+++ CHANGES.md
@@ -4,6 +4,8 @@
 2.1 (unreleased)
 ----------------
 
+- Bug 77: use `ConfigUpdater` to allow editing secrets and keeping comments.
+
 - Bug 1: provide better error message if remote user does not exist.
 
   This is also cleaning up the general error output and we're now hiding